refactor(fft): share slot cache, params ring and trig tables
Extract the helpers that were duplicated across the packed-Stockham and
packed-tiled Fourier modules into shared fourier-level modules:

- slotCache.ts: createSlotCache (was copied into three state.ts files)
- params.ts: createParamsRing (label now a parameter), dropping the two
  per-module copies
- trigTables.ts: createTrigBuffer, createFftTrigTable (generalized to
  (size, entryCount = size)) and createR2cTrigTable(windowSize)

The c2r r2c-table stays local as it uses positiveWindowSize. Cuts jscpd
clones from 8 to 4; check:ts, check:lint and the 2760 tests pass.
